Step 2: Write your first test.
Read through the code under test and pick a smallest part of functionality you
believe you understand.

Then write a test for this functionality to make sure your understanding is
correct.

If the test does not pass, figure out why. Probably there is a bug, or your
understanding of the functionality you have chosen to test is suboptimal.

By the way, this 1st test has already pointed out the bug in Database module.