Go back to Step 3: Example of non-failing mutation.
Mutations like that are a thing too. And as you could guess it does not fail,
since our test does not care about the contents of the array that we have
passed in. Mutate only things you think will break your tests.

Now we can either:
- modify our current test to cover that, or
- add another test.

I prefer the latter, since it results in more granular tests.